+++ x/ccv/provider/keeper/keeper.go
@@ -927,6 +927,19 @@ func (k Keeper) IsAllowlistEmpty(ctx sdk.Context, consumerId string) bool {
 	return !iterator.Valid()
 }
 
+// UpdateAllowlist populates the allowlist store for the consumer chain with this consumer id
+func (k Keeper) UpdateAllowlist(ctx sdk.Context, consumerId string, allowlist []string) {
+	k.DeleteAllowlist(ctx, consumerId)
+	for _, address := range allowlist {
+		consAddr, err := sdk.ConsAddressFromBech32(address)
+		if err != nil {
+			continue
+		}
+
+		k.SetAllowlist(ctx, consumerId, types.NewProviderConsAddress(consAddr))
+	}
+}
+
 // SetDenylist denylists validator with `providerAddr` address on chain `consumerId`
 func (k Keeper) SetDenylist(
 	ctx sdk.Context,
@@ -990,6 +1003,19 @@ func (k Keeper) IsDenylistEmpty(ctx sdk.Context, consumerId string) bool {
 	return !iterator.Valid()
 }
 
+// UpdateDenylist populates the denylist store for the consumer chain with this consumer id
+func (k Keeper) UpdateDenylist(ctx sdk.Context, consumerId string, denylist []string) {
+	k.DeleteDenylist(ctx, consumerId)
+	for _, address := range denylist {
+		consAddr, err := sdk.ConsAddressFromBech32(address)
+		if err != nil {
+			continue
+		}
+
+		k.SetDenylist(ctx, consumerId, types.NewProviderConsAddress(consAddr))
+	}
+}
+
 // SetMinimumPowerInTopN sets the minimum power required for a validator to be in the top N
 // for a given consumer chain.
 func (k Keeper) SetMinimumPowerInTopN(
@@ -1027,6 +1053,30 @@ func (k Keeper) DeleteMinimumPowerInTopN(
 ) {
 	store := ctx.KVStore(k.storeKey)
 	store.Delete(types.MinimumPowerInTopNKey(consumerId))
+}
+
+// UpdateMinimumPowerInTopN populates the minimum power in Top N for the consumer chain with this consumer id
+func (k Keeper) UpdateMinimumPowerInTopN(ctx sdk.Context, consumerId string, oldTopN uint32, newTopN uint32) error {
+	// if the top N changes, we need to update the new minimum power in top N
+	if newTopN != oldTopN {
+		if newTopN > 0 {
+			// if the chain receives a non-zero top N value, store the minimum power in the top N
+			bondedValidators, err := k.GetLastProviderConsensusActiveValidators(ctx)
+			if err != nil {
+				return err
+			}
+			minPower, err := k.ComputeMinPowerInTopN(ctx, bondedValidators, newTopN)
+			if err != nil {
+				return err
+			}
+			k.SetMinimumPowerInTopN(ctx, consumerId, minPower)
+		} else {
+			// if the chain receives a zero top N value, we delete the min power
+			k.DeleteMinimumPowerInTopN(ctx, consumerId)
+		}
+	}
+
+	return nil
 }
 
 func (k Keeper) UnbondingCanComplete(ctx sdk.Context, id uint64) error {
